Remove duplication of SideBar/Row
