+++ src/components/SideBar/Section.js
@@ -4,7 +4,7 @@ import classNames from 'classnames'
 function Section({children = null, style = {}}) {
     return (
         <div
-            className={classNames('sidebar-section')}
+            className="sidebar-section"
             style={style}
         >
             {children}
